+++ D3D_Shaders/stdafx.h
@@ -43,7 +43,7 @@ struct token_operand
 
 vector<DWORD> assembleIns(string s);
 vector<byte> readFile(string fileName);
-vector<DWORD> ComputeHash(byte const* input, size_t size);
+vector<DWORD> ComputeHash(byte const* input, DWORD size);
 vector<string> stringToLines(const char* start, size_t size);
 DWORD strToDWORD(string s);
 string convertF(DWORD original);
